Undo package names refactor and refactor injection via constructor
